feature: adding UI for onlineWaitingRoom and hooks
