+++ src/modules/pages/WaittingRoomComponents/hooks/useOnlineWaitingRoom.js
@@ -1,8 +1,9 @@
 import { useState, useEffect } from 'react';
 import { collection, doc, getDoc } from 'firebase/firestore';
 import { db } from '../../../../config/firebase';
+import { APP_ENV, CURRENT_DATE } from '../../../../lib/constants';
 
-const useOnlineWaitingRoom = (date) => {
+const useOnlineWaitingRoom = (date  = CURRENT_DATE) => {
     const [schedules, setSchedules] = useState([]);
     const [loading, setLoading] = useState(true);
     const [error, setError] = useState(null);
@@ -12,28 +13,36 @@ const useOnlineWaitingRoom = (date) => {
             try {
                 setLoading(true);
                 setError(null);
-
-                // Format date to ISO string to match backend format
-                const formattedDate = date.toISOString().split('T')[0];
                 
-                // Get the document from Firestore
-                const docRef = doc(db, 'doctor_schedule', formattedDate);
+                const formattedDate = date.toISOString().split('T')[0];
+                const docRef = doc(db, `${APP_ENV}_doctor_schedule`, formattedDate);
                 const docSnap = await getDoc(docRef);
-
+    
                 if (docSnap.exists()) {
                     const data = docSnap.data();
-                    setSchedules(data.schedules || []);
+                    console.log("Fetched data:", data);
+                    if (data && data.schedules && Array.isArray(data.schedules)) {
+                        data.schedules.forEach(schedule => {
+                            console.log("Time slots for schedule:", schedule.time_slots);
+                        });
+                        setSchedules(data.schedules);
+                    } else {
+                        console.log("Document exists but no valid schedules array found:", data);
+                        setSchedules([]);
+                    }
                 } else {
+                    console.log("No document found for date:", formattedDate);
                     setSchedules([]);
                 }
             } catch (err) {
                 setError(err.message);
                 console.error('Error fetching schedules:', err);
+                setSchedules([]);
             } finally {
                 setLoading(false);
             }
         };
-
+    
         if (date) {
             fetchSchedules();
         }
